+++ src/Control/Actions/SnapshotDeployAction.php
@@ -1,6 +1,5 @@
 <?php
 
-use Carbon\Carbon;
 use Slim\Http\Request;
 use Slim\Http\Response;
 use Symfony\Component\Filesystem\Filesystem;
@@ -21,9 +20,14 @@ final class SnapshotDeployAction extends AbstractAction
         if (!$fileSystem->exists($path)) {
             $this->flash->addMessage('error', $this->translator->trans('file.notexists'));
         } else {
-            $snapshotData = file_get_contents($path);
-            $this->ts->getInstance()->serverSnapshotDeploy($snapshotData, true);
-            $this->flash->addMessage('success', $this->translator->trans('done'));
+            try {
+                $snapshotData = file_get_contents($path);
+                $this->ts->getInstance()->serverSnapshotDeploy($snapshotData, true);
+                $this->flash->addMessage('success', $this->translator->trans('done'));
+            } catch (Exception $e) {
+                $this->logger->error('Could not deploy ' . $path . '. Cause: ' . $e->getMessage());
+                $this->flash->addMessage('error', $this->translator->trans('snapshots.error.deploy'));
+            }
         }
 
         return $response->withRedirect('/snapshots/' . $sid);