+++ src/Util/Auth/TSAuthAdapter.php
@@ -45,7 +45,7 @@ class TSAuthAdapter extends \Zend\Authentication\Adapter\AbstractAdapter
         $password = $this->getCredential();
 
         if ($this->ts->login($user, $password)) {
-            $this->logger->info(sprintf('Authenticated as %s', $user));
+            $this->logger->debug(sprintf('Authenticated as %s', $user));
 
             $user = ['identity' => $user, 'user' => $user, 'password'=> $password, 'role' => ACL::ACL_DEFAULT_ROLE_ADMIN];
             return new Result(Result::SUCCESS, $user, array());